Created Autonomous queue.
You add autonomous steps to it.
When you call the periodic() function, it executes steps.
When a step says it is done, it is removed from the queue and finish()
is called on it.
The next one has start() called on it.
When a step says it is not done, it is left in the queue until it is
done.
Also, you can clear() the queue.